Fixes #38073 - Support Rails 7.0
