+++ app/lib/actions/foreman_leapp/preupgrade_job.rb
@@ -8,8 +8,8 @@ module Actions
       end
 
       def plan(job_invocation, host, *_args)
-        return unless ::JobHelper.correct_feature?(job_invocation, 'leapp_preupgrade') ||
-                      ::JobHelper.correct_feature?(job_invocation, 'leapp_remediation_plan')
+        return unless ::Helpers::JobHelper.correct_feature?(job_invocation, 'leapp_preupgrade') ||
+                      ::Helpers::JobHelper.correct_feature?(job_invocation, 'leapp_remediation_plan')
 
         plan_self(host_id: host.id, job_invocation_id: job_invocation.id)
       end
